requirements.txt
Django==5.2.5
python-dotenv==1.1.1
djangorestframework==3.16.1
django-autoslug==1.9.9
django-countries==7.6.1
django-phonenumber-field==8.1.0
phonenumbers==9.0.13
Pillow==11.3.0
argon2-cffi==25.1.0
django-taggit==6.1.0
django-filter==25.1
black==25.1.0
django-allauth==65.11.1
Faker==30.8.2
psycopg2-binary==2.9.10

# Development dependencies (optional - install only if needed)
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.18.1
